move GEMDigiProducer to plugin
