+++ SimMuon/GEMDigitizer/plugins/BuildFile.xml
@@ -1,16 +1,16 @@
+<use name="FWCore/Utilities"/>
+<use name="FWCore/Framework"/>
+<use name="FWCore/ServiceRegistry"/>
+<use name="CommonTools/UtilAlgos"/>
+<use name="CondFormats/RecoMuonObjects"/>
+<use name="DataFormats/Common"/>
+<use name="DataFormats/MuonDetId"/>
+<use name="DataFormats/GeometryVector"/>
+<use name="Geometry/Records"/>
+<use name="CondFormats/DataRecord"/>
+<use name="SimGeneral/PreMixingModule"/>
+<use name="SimMuon/GEMDigitizer"/>
 
 <library file="*.cc" name="SimMuonGEMDigitizerPlugins"> 
-  <use name="FWCore/Utilities"/>
-  <use name="FWCore/Framework"/>
-  <use name="FWCore/ServiceRegistry"/>
-  <use name="CommonTools/UtilAlgos"/>
-  <use name="CondFormats/RecoMuonObjects"/>
-  <use name="DataFormats/Common"/>
-  <use name="DataFormats/MuonDetId"/>
-  <use name="DataFormats/GeometryVector"/>
-  <use name="Geometry/Records"/>
-  <use name="CondFormats/DataRecord"/>
-  <use name="SimGeneral/PreMixingModule"/>
-
   <flags EDM_PLUGIN="1"/>
 </library>
