+++ SimMuon/GEMDigitizer/plugins/GEMDigiProducer.h
@@ -42,4 +42,3 @@ private:
 };
 
 #endif
-
