+++ SimMuon/GEMDigitizer/src/SealModule.cc
@@ -3,8 +3,8 @@
 #include "SimMuon/GEMDigitizer/interface/ME0DigiPreRecoModelFactory.h"
 #include "SimMuon/GEMDigitizer/interface/ME0DigiModelFactory.h"
 
-#include "SimMuon/GEMDigitizer/interface/GEMDigiProducer.h"
-DEFINE_FWK_MODULE(GEMDigiProducer);
+//#include "SimMuon/GEMDigitizer/interface/GEMDigiProducer.h"
+//DEFINE_FWK_MODULE(GEMDigiProducer);
 
 #include "SimMuon/GEMDigitizer/interface/GEMPadDigiProducer.h"
 DEFINE_FWK_MODULE(GEMPadDigiProducer);
